exposure time for presets are now in float ref Feature #343

+++ leginon/leginondata.py
@@ -457,7 +457,7 @@ class PresetData(InSessionData):
 			('dimension', dict),
 			('binning', dict),
 			('offset', dict),
-			('exposure time', int),
+			('exposure time', float),
 			('removed', bool),
 			('hasref', bool),
 			('dose', float),
